% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{duckhts_bench_cache_path}
\alias{duckhts_bench_cache_path}
\title{Return a path below the configured DuckHTS cache.}
\usage{
duckhts_bench_cache_path(relative_path = "")
}
\arguments{
\item{relative_path}{Relative cache path.}
}
\value{
A cache path.
}
\description{
Return a path below the configured DuckHTS cache.
}
